+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/dto/response/RevenueDTO.java
@@ -1,0 +1,16 @@
+package com.ivanovp.medical_record.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class RevenueDTO {
+
+    private String label;
+    private BigDecimal revenue;
+}